requirements.txt
beautifulsoup4==4.9.3
blis==0.4.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
cymem==2.0.3
fr-core-news-sm==2.3.0
geographiclib==1.50
geopy==2.0.0
idna==2.10
murmurhash==1.0.2
mysql-connector-python==8.0.22
numpy==1.19.3
plac==1.1.3
preshed==3.0.2
protobuf==3.13.0
python-dotenv==0.15.0
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
spacy==2.3.2
srsly==1.0.2
thinc==7.4.1
tqdm==4.51.0
urllib3==1.25.11
wasabi==0.8.0
